crud fully parallelized, runners re-organized
