+++ example_app/example_app/settings.py
@@ -96,10 +96,10 @@ WSGI_APPLICATION = "example_app.wsgi.application"
 # PostgreSQL Database Configuration
 DATABASE_CONFIGURATION = {
     "ENGINE": "django.db.backends.postgresql",
-    "NAME": environ.get("POSTGRES_DB_NAME"),
-    "USER": environ.get("POSTGRES_USER"),
-    "PASSWORD": environ.get("POSTGRES_PASSWORD"),
-    "HOST": environ.get("POSTGRES_HOST"),
+    "NAME": environ.get("DB_NAME"),
+    "USER": environ.get("DB_USER"),
+    "PASSWORD": environ.get("DB_PASS"),
+    "HOST": environ.get("DB_HOST"),
     "PORT": "5432",
 }
