+++ python/noCRUD.py
@@ -4,17 +4,13 @@ import importlib.util
 import inspect
 from typing import Dict, Callable
 import time
-import traceback
-import requests
-import io
-from contextlib import redirect_stdout
-from multiprocessing import Pool
 
 ### Local Utils
 from flows.biz_logic import pitch_lock_after_interactions
-from utils.provisioning import provision_env_for_flow
+from runners.main import crud_flows_runner
+from runners.serial import request_flows_runner
 from utils.db_client import DBClient
-from utils.printing import format_crud_print, print_group_separator, print_warn
+from utils.printing import print_group_separator, print_warn
 from utils.decorators import with_stack_trace
 
 ##### Import Manually Registered Flows
@@ -118,15 +114,15 @@ def main():
     # "Main"
     #####################################
 
-    # Run Migrations just to ensure we have a clean env
+    #  -- This should only be on the serial side -- Run Migrations just to ensure we have a clean env
     start_time = time.perf_counter()
     print_group_separator("Initial Setup")
-    db = DBClient()
-    db.reset()
 
     is_parallel = True
     if args.serial:
         is_parallel = False
+        db = DBClient()
+        db.reset()
 
     # Filter and Run
     print_group_separator("Run Flows")
@@ -140,7 +136,6 @@ def main():
     if args.crud:
         flows_to_run = CRUD_FLOWS.keys()
         print(f"Flows to run: {flows_to_run}")
-        print(is_parallel)
         crud_flows_runner(flows_to_run, allFlows, parallel=is_parallel)
 
     if args.flows:
@@ -151,111 +146,6 @@ def main():
     end_time = time.perf_counter()
     print("=" * 80)
     print(f"\nTest runner took: {end_time - start_time:.6f} seconds")
-
-
-def run_isolated_flow(flow_name_and_func):
-    flow_name, flow_function = flow_name_and_func
-
-    provision_env_for_flow(flow_name)
-
-    buffer = io.StringIO()
-    try:
-        with redirect_stdout(buffer):
-            print(f"Running flow: {flow_name}\n{'-' * 60}")
-            res = flow_function()
-            formatted = format_crud_print(res)
-            print("\n")
-            print(formatted)
-            print("\n")
-        return flow_name, formatted, buffer.getvalue()
-    except requests.exceptions.ConnectionError:
-        msg = (
-            f"\nFlow '{flow_name}' failed: Unable to connect to the backend. "
-            f"Is it running?\n"
-        )
-        return flow_name, msg.strip(), buffer.getvalue() + msg
-    except Exception as e:
-        tb = traceback.format_exc()
-        err = f"\nFlow '{flow_name}' failed with error: {e}\nStack trace:\n{tb}"
-        return flow_name, f"Fail: {e}", buffer.getvalue() + err
-
-
-def crud_flows_runner(flowsToRun, allFlows, parallel=True):
-    """Run Flows with output that expects CRUD hashmap returned from each flow"""
-
-    # Pair flow names with functions so we can parallelize
-    flow_items = [(name, allFlows[name]) for name in flowsToRun]
-
-    if parallel:
-        with Pool() as pool:
-            results = pool.map(run_isolated_flow, flow_items)
-    else:
-        results = crud_flows_runner_serial(flowsToRun, allFlows)
-
-    # Output and summary collection
-    crud_results = {}
-    for flow_name, formatted_result, captured_output in results:
-        print(captured_output)
-        crud_results[flow_name] = formatted_result
-
-    # Print test case Results
-    print_group_separator("Results Summary")
-    max_length = max(len(k) for k in crud_results.keys())
-    for k, v in crud_results.items():
-        print(f"{k.ljust(max_length)}: {v}")
-
-
-def crud_flows_runner_serial(flowsToRun, allFlows):
-    """Run Flows with output that expects CRUD results returned from each flow"""
-    results = []
-    for flow_name in flowsToRun:
-        flow_function = allFlows[flow_name]
-        buffer = io.StringIO()
-        try:
-            with redirect_stdout(buffer):
-                print(f"Running flow: {flow_name}\n{'-' * 60}")
-                res = flow_function()
-                formatted = format_crud_print(res)
-                print("\n")
-                print(formatted)
-                print("\n")
-            results.append((flow_name, formatted, buffer.getvalue()))
-        except requests.exceptions.ConnectionError:
-            msg = f"\nFlow '{flow_name}' failed: Unable to connect to the backend. Is it running?\n"
-            results.append((flow_name, msg.strip(), buffer.getvalue() + msg))
-        except Exception as e:
-            tb = traceback.format_exc()
-            err = f"\nFlow '{flow_name}' failed with error: {e}\nStack trace:\n{tb}"
-            results.append((flow_name, f"Fail: {e}", buffer.getvalue() + err))
-    return results
-
-
-def request_flows_runner(flowstoRun, allFlows):
-    """Run Flows with the standard request flows style output"""
-    test_case_results = {}
-    for flow_name in flowstoRun:
-        flow_function = allFlows[flow_name]
-        try:
-            print(f"Running flow: {flow_name}\n{'-' * 60}")
-            test_case_results[flow_name] = flow_function()
-            print(f"\nFlow '{flow_name}' completed successfully. \n")
-        except requests.exceptions.ConnectionError:
-            print(
-                f"\nFlow '{flow_name}' failed: Unable to connect to the backend. Is it running? \n"
-            )
-        except Exception as e:
-            test_case_results[flow_name] = f"Fail: {e}"
-            print(f"\nFlow '{flow_name}' failed with error: {e} \n")
-            print("Stack trace:")
-            print(traceback.format_exc())
-
-    # Print test case Results
-    print_group_separator("Results Summary")
-    # Determine the max length for entity names to align columns properly
-    max_length = max(len(k) for k in test_case_results.keys())
-
-    for k, v in test_case_results.items():
-        print(f"{k.ljust(max_length)}: {v}")
 
 
 @with_stack_trace
