+++ python/utils/crud.py
@@ -10,7 +10,6 @@ def simple_create(
     """Assumes that the object is not dependent on other obejcts (and needs no other modification before request is sent)"""
     obj = get_fixture_by_index(fixtureName, fixtureIndex)
     res = api.create_object(endpoint, obj)
-    print("res", res)
     return res[idField]
