+++ python/utils/db_client.py
@@ -145,7 +145,6 @@ class DBClient:
         subprocess.run(
             ["python", "manage.py", "migrate"], cwd=app_dir, check=True
         )  # Python subprocesses auto block so no need to use process.wait()
-        print("app_dir", app_dir)
         addFixtures()
         return
